Change LessonList to an observable list
Previously, payment status could not update when a new lesson was added, making
payment status immutable for the entire month.

Hence, we will need a listener in LessonList to listen for any changes in
adding/deleting lessons, as this will change the total amount of the payment
for the yearmonth in the student's payment list.