Fix ModelManager tests
Previously, model manager tests kept failing with actual values being different
to the payments allowed. This was because in each Student Builder, a default
lesson list was being created, which is deeply integrated into the logic
of calculating total earnings.

Hence, we shall create a StudentStub to create temporary dummy methods
that only test the listeners without the lesson logic. This allows
us to test the recomputeTotal methods in isolation.